[SKIP]: Create `git-read` GitHub Action
Utilize the [`seanpdoyle/git-read`][git-read] [GitHub Action] to
generate web pages from this project's Git history.

The declarations in
[.github/workflow/git-read.yml](#.github/workflow/git-ready.yml) were
inspired by the corresponding GitHub Action workflow declared in the
[`git-read` repository][git-read-workflow].

Specifying the `--git-dir` and `--output-dir` values in terms of the
GitHub Actions-provided [`$GITHUB_WORKSPACE` environment variable][env]
is important to ensure that the proper output is pushed to the
`gh-pages` branch.

Upon each `git push` to this repository, this project's Git history will
be rendered as HTML and deployed to the [`gh-pages` branch][gh-pages],
available at https://seanpdoyle.github.io/content-authoring.

[git-read]: https://github.com/seanpdoyle/git-read
[gh-action]: https://github.com/features/actions
[git-read-workflow]: https://github.com/seanpdoyle/git-read/blob/main/.github/workflows/git-read.yml
[env]: https://help.github.com/en/actions/configuring-and-managing-workflows/using-environment-variables#default-environment-variables
[gh-pages]: https://pages.github.com/